// frontend/src/lib/api/models/videoAdminUpdateRequestContentPrivacy.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Stream Archive API
 * 멀티 플랫폼 스트리밍 녹화 시스템 API
 * OpenAPI spec version: 1.0.0
 */

/**
 * 콘텐츠 공개 범위 (PUBLIC/UNLISTED/PRIVATE)
 */
export type VideoAdminUpdateRequestContentPrivacy =
  (typeof VideoAdminUpdateRequestContentPrivacy)[keyof typeof VideoAdminUpdateRequestContentPrivacy];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VideoAdminUpdateRequestContentPrivacy = {
  PUBLIC: "PUBLIC",
  UNLISTED: "UNLISTED",
  PRIVATE: "PRIVATE",
} as const;